When changing views, prefer to give focus to clients whose tags exactly match the tags being viewed (rather than a proper subset), and fall back to the old behaviour if no such clients exist.
The old behaviour is annoying when I've got a single client which is
visible under every tag. If such a client exists, it will always get
focus when switching views, rather than the last selected client of that
view. I generally run vlc this way, and I obviously don't want to
interact with it much.